some stuff?

+++ tenmo-server/src/main/java/com/techelevator/tenmo/model/Account.java
@@ -4,25 +4,25 @@ import java.math.BigDecimal;
 
 public class Account {
 
-    private Integer accountId;
-    private Integer userID;
+    private Long accountId;
+    private Long userID;
     private BigDecimal balance;
 
 
 
-    public Integer getAccountId() {
+    public Long getAccountId() {
         return accountId;
     }
 
-    public void setAccountId(Integer accountId) {
+    public void setAccountId(Long accountId) {
         this.accountId = accountId;
     }
 
-    public Integer getUserID() {
+    public Long getUserID() {
         return userID;
     }
 
-    public void setUserID(Integer userID) {
+    public void setUserID(Long userID) {
         this.userID = userID;
     }
